# openapi/openapi.yaml
openapi: 3.1.0
info: {title: OpenAPI definition, version: v0}
servers:
- {url: 'http://localhost:8080', description: Generated server url}
paths:
...
  /api/applications:
    put:
      tags: [application-resource]
      operationId: updateApplication
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateApplicationDTO'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationForApplicantDTO'}
  /api/applications/all:
    get:
      tags: [application-resource]
      operationId: getAllApplications
      parameters:
      - name: pageSize
        in: query
        required: false
        schema: {type: integer, format: int32, minimum: 1}
      - name: pageNumber
        in: query
        required: false
        schema: {type: integer, format: int32, minimum: 0}
      - name: states
        in: query
        required: false
        schema:
          type: array
          items: {type: string}
      - name: researchGroupIds
        in: query
        required: false
        schema:
          type: array
          items: {type: string, format: uuid}
      - name: supervisingProfessorIds
        in: query
        required: false
        schema:
          type: array
          items: {type: string, format: uuid}
      - name: jobIds
        in: query
        required: false
        schema:
          type: array
          items: {type: string, format: uuid}
      - name: sortBy
        in: query
        required: false
        schema: {type: string}
      - name: direction
        in: query
        required: false
        schema:
          type: string
          enum: [ASC, DESC]
      - name: searchQuery
        in: query
        required: false
        schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageAdminApplicationOverviewDTO'}
  /api/applications/create/{jobId}:
    post:
      tags: [application-resource]
      operationId: createApplication
      parameters:
      - name: jobId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationForApplicantDTO'}
  /api/applications/documents/{documentId}:
    delete:
      tags: [application-resource]
      operationId: deleteDocumentFromApplication
      parameters:
      - name: documentId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: OK}
  /api/applications/documents/{documentId}/name:
    put:
      tags: [application-resource]
      operationId: renameDocument
      parameters:
      - name: documentId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: newName
        in: query
        required: true
        schema: {type: string}
      responses:
        '200': {description: OK}
  /api/applications/getDocumentIds/{applicationId}:
    get:
      tags: [application-resource]
      operationId: getDocumentIds
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationDocumentIdsDTO'}
  /api/applications/pages:
    get:
      tags: [application-resource]
      operationId: getApplicationPages
      parameters:
      - name: pageSize
        in: query
        required: false
        schema: {type: integer, format: int32, minimum: 1}
      - name: pageNumber
        in: query
        required: false
        schema: {type: integer, format: int32, minimum: 0}
      - name: sortBy
        in: query
        required: false
        schema: {type: string}
      - name: direction
        in: query
        required: false
        schema:
          type: string
          enum: [ASC, DESC]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PageApplicationOverviewDTO'}
  /api/applications/withdraw/{applicationId}:
    put:
      tags: [application-resource]
      operationId: withdrawApplication
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: OK}
  /api/applications/{applicationId}:
    get:
      tags: [application-resource]
      operationId: getApplicationById
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationForApplicantDTO'}
    delete:
      tags: [application-resource]
      operationId: deleteApplication
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: OK}
...
  /api/applications/{applicationId}/detail:
    get:
      tags: [application-resource]
      operationId: getApplicationForDetailPage
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplicationDetailDTO'}
  /api/applications/{applicationId}/documents/{documentType}:
    post:
      tags: [application-resource]
      summary: Upload documents
      operationId: uploadDocuments
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: documentType
        in: path
        required: true
        schema:
          type: string
          enum: [BACHELOR_TRANSCRIPT, MASTER_TRANSCRIPT, REFERENCE, REFERENCE_LETTER,
            CV, CUSTOM]
      requestBody:
        content:
          multipart/form-data:
            schema: {$ref: '#/components/schemas/MultipartUploadRequest'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/DocumentInformationHolderDTO'}
                uniqueItems: true
...
  /api/applications/{applicationId}/references:
    get:
      tags: [reference-request-resource]
      operationId: getReferences
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ReferenceRequestDTO'}
    post:
      tags: [reference-request-resource]
      operationId: add
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefereeContactDTO'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReferenceRequestDTO'}
  /api/applications/{applicationId}/references/{referenceId}:
    put:
      tags: [reference-request-resource]
      operationId: update
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: referenceId
        in: path
        required: true
        schema: {type: string, format: uuid}
      requestBody:
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefereeContactDTO'}
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReferenceRequestDTO'}
    delete:
      tags: [reference-request-resource]
      operationId: remove
      parameters:
      - name: applicationId
        in: path
        required: true
        schema: {type: string, format: uuid}
      - name: referenceId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: OK}
...
  /api/documents/{documentId}:
    get:
      tags: [document-resource]
      operationId: downloadDocument
      parameters:
      - name: documentId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {type: string, format: binary}
    delete:
      tags: [document-resource]
      operationId: deleteDocument
      parameters:
      - name: documentId
        in: path
        required: true
        schema: {type: string, format: uuid}
      responses:
        '200': {description: OK}
...
components:
  schemas:
...
    AcquaintanceDepth:
      type: string
      enum: [CASUALLY, MODERATELY, WELL, VERY_WELL]
    AcquaintanceDuration:
      type: string
      enum: [LESS_THAN_ONE_YEAR, ONE_TO_TWO_YEARS, THREE_TO_FIVE_YEARS, MORE_THAN_FIVE_YEARS]
...
    AdminApplicationOverviewDTO:
      type: object
      properties:
        applicantAvatar: {type: string}
        applicantName: {type: string}
        applicantUserId: {type: string, format: uuid}
        applicationId: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
        jobId: {type: string, format: uuid}
        jobTitle: {type: string}
        researchGroupId: {type: string, format: uuid}
        researchGroupName: {type: string}
        state:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        supervisingProfessorId: {type: string, format: uuid}
        supervisingProfessorName: {type: string}
      required: [applicantUserId, applicationId, jobId]
...
    ApplicantDTO:
      type: object
      properties:
        bachelorDegreeName: {type: string}
        bachelorGrade: {type: string}
        bachelorGradeLowerLimit: {type: string}
        bachelorGradeUpperLimit: {type: string}
        bachelorUniversity: {type: string}
        city: {type: string}
        country: {type: string}
        masterDegreeName: {type: string}
        masterGrade: {type: string}
        masterGradeLowerLimit: {type: string}
        masterGradeUpperLimit: {type: string}
        masterUniversity: {type: string}
        postalCode: {type: string}
        street: {type: string}
        user: {$ref: '#/components/schemas/UserDTO'}
      required: [user]
    ApplicantForApplicationDetailDTO:
      type: object
      properties:
        bachelorDegreeName: {type: string}
        bachelorGrade: {type: string}
        bachelorGradeLowerLimit: {type: string}
        bachelorGradeUpperLimit: {type: string}
        bachelorUniversity: {type: string}
        city: {type: string}
        country: {type: string}
        masterDegreeName: {type: string}
        masterGrade: {type: string}
        masterGradeLowerLimit: {type: string}
        masterGradeUpperLimit: {type: string}
        masterUniversity: {type: string}
        postalCode: {type: string}
        street: {type: string}
        user: {$ref: '#/components/schemas/UserForApplicationDetailDTO'}
      required: [user]
    ApplicationDetailDTO:
      type: object
      properties:
        applicant: {$ref: '#/components/schemas/ApplicantForApplicationDetailDTO'}
        applicationId: {type: string, format: uuid}
        applicationState:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        desiredDate: {type: string, format: date}
        jobEndDate: {type: string, format: date}
        jobId: {type: string, format: uuid}
        jobLocation:
          type: string
          enum: [GARCHING, GARCHING_HOCHBRUECK, HEILBRONN, MUNICH, STRAUBING, WEIHENSTEPHAN,
            SINGAPORE]
        jobTitle: {type: string}
        motivation: {type: string}
        projects: {type: string}
        recommendationType: {$ref: '#/components/schemas/RecommendationType'}
        referenceLettersConfidential: {type: boolean}
        referenceLettersRequired: {type: integer, format: int32}
        references:
          type: array
          items: {$ref: '#/components/schemas/ReferenceRequestDTO'}
        researchGroup: {type: string}
        specialSkills: {type: string}
        supervisingProfessorName: {type: string}
      required: [applicationId, applicationState, jobId, researchGroup, supervisingProfessorName]
    ApplicationDocumentIdsDTO:
      type: object
      properties:
        bachelorDocumentIds:
          type: array
          items: {$ref: '#/components/schemas/DocumentInformationHolderDTO'}
          uniqueItems: true
        cvDocumentId: {$ref: '#/components/schemas/DocumentInformationHolderDTO'}
        masterDocumentIds:
          type: array
          items: {$ref: '#/components/schemas/DocumentInformationHolderDTO'}
          uniqueItems: true
        referenceDocumentIds:
          type: array
          items: {$ref: '#/components/schemas/DocumentInformationHolderDTO'}
          uniqueItems: true
...
    ApplicationForApplicantDTO:
      type: object
      properties:
        applicant: {$ref: '#/components/schemas/ApplicantDTO'}
        applicationId: {type: string, format: uuid}
        applicationState:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        desiredDate: {type: string, format: date}
        job: {$ref: '#/components/schemas/JobCardDTO'}
        motivation: {type: string}
        projects: {type: string}
        referenceLettersConfidential: {type: boolean}
        references:
          type: array
          items: {$ref: '#/components/schemas/ReferenceRequestDTO'}
        specialSkills: {type: string}
      required: [applicationState, job]
    ApplicationOverviewDTO:
      type: object
      properties:
        applicationId: {type: string, format: uuid}
        applicationState:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        createdAt: {type: string, format: date-time}
        jobId: {type: string, format: uuid}
        jobTitle: {type: string}
        recommendationMissing: {type: boolean}
        researchGroup: {type: string}
...
    DocumentInformationHolderDTO:
      type: object
      properties:
        documentType:
          type: string
          enum: [BACHELOR_TRANSCRIPT, MASTER_TRANSCRIPT, REFERENCE, REFERENCE_LETTER,
            CV, CUSTOM]
        id: {type: string, format: uuid}
        name: {type: string}
        size: {type: integer, format: int64}
      required: [id, size]
...
    JobCardDTO:
      type: object
      properties:
        applicationId: {type: string, format: uuid}
        applicationState:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        avatar: {type: string}
        contractDuration: {type: integer, format: int32}
        imageUrl: {type: string}
        jobId: {type: string, format: uuid}
        location:
          type: string
          enum: [GARCHING, GARCHING_HOCHBRUECK, HEILBRONN, MUNICH, STRAUBING, WEIHENSTEPHAN,
            SINGAPORE]
        professorName: {type: string}
        recommendationType: {$ref: '#/components/schemas/RecommendationType'}
        referenceLettersRequired: {type: integer, format: int32}
        relativeTimeEnglish: {type: string}
        relativeTimeGerman: {type: string}
        startDate: {type: string, format: date}
        subjectArea:
          type: string
          enum: [AEROSPACE_ENGINEERING, AGRICULTURAL_ENGINEERING, AGRICULTURAL_SCIENCE,
            ARCHITECTURE, ART_HISTORY, AUTOMOTIVE_ENGINEERING, BIOENGINEERING, BIOCHEMISTRY,
            BIOLOGY, BIOMEDICAL_ENGINEERING, BIOTECHNOLOGY, CHEMISTRY, COMPUTER_ENGINEERING,
            COMPUTER_SCIENCE, COMPUTER_VISION, DATA_SCIENCE, ECONOMICS, EDUCATION_TECHNOLOGY,
            ELECTRICAL_ENGINEERING, ENERGY_SYSTEMS, ENVIRONMENTAL_BIOLOGY, ENVIRONMENTAL_CHEMISTRY,
            ENVIRONMENTAL_ENGINEERING, ENVIRONMENTAL_LAW, ENVIRONMENTAL_SCIENCE, FINANCIAL_ENGINEERING,
            FOOD_TECHNOLOGY, GEOLOGY, GEOSCIENCES, INDUSTRIAL_ENGINEERING, INFORMATION_SYSTEMS,
            LIFE_SCIENCES, LINGUISTICS, MARINE_BIOLOGY, MATERIALS_SCIENCE, MATHEMATICS,
            MECHANICAL_ENGINEERING, MEDICAL_INFORMATICS, NEUROSCIENCE, PHILOSOPHY,
            PHYSICS, PSYCHOLOGY, SOFTWARE_ENGINEERING, SPORTS_SCIENCE, STATISTICS,
            TELECOMMUNICATIONS, URBAN_PLANNING]
        title: {type: string}
        workload: {type: integer, format: int32}
      required: [jobId, location, professorName, subjectArea, title]
...
    MultipartUploadRequest:
      type: object
      properties:
        files: {type: string, format: binary, description: List of documents to upload}
...
    OverallRecommendation:
      type: string
      enum: [HIGHEST_ENTHUSIASM, STRONGLY_RECOMMEND, RECOMMEND, RECOMMEND_WITH_RESERVATIONS,
        DO_NOT_RECOMMEND]
    PageAdminApplicationOverviewDTO:
      type: object
      properties:
        content:
          type: array
          items: {$ref: '#/components/schemas/AdminApplicationOverviewDTO'}
        empty: {type: boolean}
        first: {type: boolean}
        last: {type: boolean}
        number: {type: integer, format: int32}
        numberOfElements: {type: integer, format: int32}
        pageable: {$ref: '#/components/schemas/PageableObject'}
        size: {type: integer, format: int32}
        sort: {$ref: '#/components/schemas/SortObject'}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer, format: int32}
...
    PageApplicationOverviewDTO:
      type: object
      properties:
        content:
          type: array
          items: {$ref: '#/components/schemas/ApplicationOverviewDTO'}
        empty: {type: boolean}
        first: {type: boolean}
        last: {type: boolean}
        number: {type: integer, format: int32}
        numberOfElements: {type: integer, format: int32}
        pageable: {$ref: '#/components/schemas/PageableObject'}
        size: {type: integer, format: int32}
        sort: {$ref: '#/components/schemas/SortObject'}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer, format: int32}
...
    PageableObject:
      type: object
      properties:
        offset: {type: integer, format: int64}
        pageNumber: {type: integer, format: int32}
        pageSize: {type: integer, format: int32}
        paged: {type: boolean}
        sort: {$ref: '#/components/schemas/SortObject'}
        unpaged: {type: boolean}
...
    PeerRating:
      type: string
      enum: [TOP_ONE_TO_TWO_PERCENT, TOP_FIVE_PERCENT, TOP_TEN_PERCENT, TOP_TWENTY_FIVE_PERCENT,
        TOP_FIFTY_PERCENT, BELOW_AVERAGE, CANNOT_JUDGE]
...
    RecommendationType:
      type: string
      enum: [LETTER_ONLY, EVALUATION_ONLY, LETTER_AND_EVALUATION]
    RefereeContactDTO:
      type: object
      properties:
        email: {type: string, format: email, maxLength: 320, minLength: 0}
        firstName: {type: string, maxLength: 255, minLength: 0}
        lastName: {type: string, maxLength: 255, minLength: 0}
        title: {type: string, maxLength: 32, minLength: 0}
      required: [email, firstName, lastName]
    RefereeRelationship:
      type: string
      enum: [COURSE_INSTRUCTOR, RESEARCH_SUPERVISOR, THESIS_ADVISOR, EMPLOYER, ACADEMIC_ADVISOR,
        OTHER]
...
    ReferenceRequestDTO:
      type: object
      properties:
        acquaintanceDepth: {$ref: '#/components/schemas/AcquaintanceDepth'}
        acquaintanceDuration: {$ref: '#/components/schemas/AcquaintanceDuration'}
        deadline: {type: string, format: date-time}
        documentId: {type: string, format: uuid}
        email: {type: string}
        firstName: {type: string}
        lastName: {type: string}
        overallRecommendation: {$ref: '#/components/schemas/OverallRecommendation'}
        ratingCollaboration: {$ref: '#/components/schemas/PeerRating'}
        ratingCommunication: {$ref: '#/components/schemas/PeerRating'}
        ratingIntellectualAbility: {$ref: '#/components/schemas/PeerRating'}
        ratingLeadership: {$ref: '#/components/schemas/PeerRating'}
        ratingMotivation: {$ref: '#/components/schemas/PeerRating'}
        ratingResearchPotential: {$ref: '#/components/schemas/PeerRating'}
        referenceRequestId: {type: string, format: uuid}
        relationship: {$ref: '#/components/schemas/RefereeRelationship'}
        status:
          type: string
          enum: [ADDED, REQUESTED, SUBMITTED, EXPIRED, DECLINED, CANCELLED]
        title: {type: string}
...
    SortObject:
      type: object
      properties:
        empty: {type: boolean}
        sorted: {type: boolean}
        unsorted: {type: boolean}
...
    UpdateApplicationDTO:
      type: object
      properties:
        applicant: {$ref: '#/components/schemas/ApplicantDTO'}
        applicationId: {type: string, format: uuid}
        applicationState:
          type: string
          enum: [SAVED, SENT, ACCEPTED, IN_REVIEW, REJECTED, WITHDRAWN, JOB_CLOSED,
            JOB_CLOSED_DRAFT, INTERVIEW]
        desiredDate: {type: string, format: date}
        motivation: {type: string}
        projects: {type: string}
        referenceLettersConfidential: {type: boolean}
        specialSkills: {type: string}
      required: [applicant, applicationId, applicationState]
...